::: MILESTONE 2 :::
Fixing the bugs.
1. Fixed the bug in saving precomputed logarithmic weight in the
docWeights.bin file
2. Modified the Controller so that it intializes all the indexes at the
time of prompting the user for directory. Previously the indexes were
loaded for each new query being searched which increases the query
processing time

+++ src/com/csulb/edu/set/query/QueryRunner.java
@@ -84,7 +84,7 @@ public class QueryRunner {
 			List<PositionalPosting> termPostingsList = diskPositionalIndex.getPostings(PorterStemmer.processToken(term));
 			 
 			 // Calculate wqt for this term
-			 double wqt = Math.log((1 + (corpusSize / termPostingsList.size())));
+			 double wqt = Math.log((1 + ((double) corpusSize / termPostingsList.size())));
 			 
 			 for (PositionalPosting pPosting : termPostingsList) {
 				 double newScore = 0;
